-document.tsx + AppRegistry.getApplication() TS support
